pyproject.toml
[project]
name = "isgamemodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.17.8",
    "matplotlib>=3.10.7",
    "mss>=10.1.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
